chore(tests): move foundation tests
