+++ tests/Foundation/Security/Token/TokenManagerTest.php
@@ -1,0 +1,49 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Foundation\Security\Token;
+
+use App\Foundation\Security\Token\TokenManagerInterface;
+use App\Foundation\ValueObject\Email;
+use App\Siklid\Document\RefreshToken;
+use App\Siklid\Document\User;
+use App\Tests\Concern\KernelTestCaseTrait;
+use App\Tests\Concern\Util\WithFaker;
+use App\Tests\TestCase;
+
+/**
+ * @psalm-suppress MissingConstructor
+ */
+class TokenManagerTest extends TestCase
+{
+    use KernelTestCaseTrait;
+    use WithFaker;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->touchCollection(RefreshToken::class);
+    }
+
+    /**
+     * @test
+     *
+     * @psalm-suppress MixedAssignment The faker email is a string
+     * @psalm-suppress MixedArgument The faker email is a string
+     */
+    public function create_access_token(): void
+    {
+        $container = $this->container();
+        $user = new User();
+        $email = $this->faker->email();
+        $user->setEmail(Email::fromString($email));
+        $sut = $container->get(TokenManagerInterface::class);
+
+        $accessToken = $sut->createAccessToken($user);
+
+        $this->assertNotNull($accessToken->getToken());
+        $this->assertExists(RefreshToken::class, ['username' => $user->getUserIdentifier()]);
+    }
+}
